# Conflicts:
#	scripts/pipeline-launch/launch.sh

+++ scripts/pipeline-launch/launch.sh
@@ -108,7 +108,7 @@ function install_pip_package {
      if [ "$_DOWNLOAD_RESULT" -ne 0 ];
         then
             echo "[ERROR] ${_DIST_NAME} download failed. Exiting"
-            exit "$_DOWNLOAD_RESULT"
+            exit_init "$_DOWNLOAD_RESULT"
         fi
     $CP_PYTHON2_PATH -m pip install $CP_PIP_EXTRA_ARGS ${_DIST_NAME}.tar.gz -q -I
     _INSTALL_RESULT=$?
@@ -116,7 +116,7 @@ function install_pip_package {
     if [ "$_INSTALL_RESULT" -ne 0 ];
     then
         echo "[ERROR] Failed to install ${_DIST_NAME}. Exiting"
-        exit "$_INSTALL_RESULT"
+        exit_init "$_INSTALL_RESULT"
     fi
 
 
@@ -133,7 +133,7 @@ function mount_nfs_if_required {
             if [ "$_SETUP_RESULT" -ne 0 ];
             then
                 echo "[ERROR] Mounting NFS failed. Exiting"
-                exit "$_SETUP_RESULT"
+                exit_init "$_SETUP_RESULT"
             fi
      fi
 
@@ -154,7 +154,7 @@ function setup_nfs_if_required {
           if [ "$_SETUP_RESULT" -ne 0 ];
           then
                 echo "[ERROR] NFS mount failed. Exiting"
-                exit "$_SETUP_RESULT"
+                exit_init "$_SETUP_RESULT"
           fi
           echo "------"
     fi
@@ -298,7 +298,7 @@ function cp_cap_init {
             if [ $_CAP_INIT_SCRIPT_RESULT -ne 0 ];
             then
                   echo "[ERROR] $_CAP_INIT_SCRIPT failed with $_CAP_INIT_SCRIPT_RESULT. Exiting"
-                  exit "$_CAP_INIT_SCRIPT_RESULT"
+                  exit_init "$_CAP_INIT_SCRIPT_RESULT"
             else
                   # Reload env vars, in case they were updated within cap init scripts
                   source "$CP_ENV_FILE_TO_SOURCE"
@@ -923,6 +923,76 @@ function self_terminate_on_cleanup_timeout() {
       fi
 }
 
+function exit_stage() {
+    local _EXIT_CODE="$1"
+    local _RECOVERY_MODE="$2"
+
+    if [ "$_RECOVERY_MODE" == "continue" ]; then
+        tag_run "$CP_CAP_RECOVERY_TAG" "true"
+        echo "Recovering from $_EXIT_CODE using continue recovery mode..."
+        return "$_EXIT_CODE"
+    fi
+    if [ "$_RECOVERY_MODE" == "sleep" ]; then
+        tag_run "$CP_CAP_RECOVERY_TAG" "true"
+        echo "Recovering from $_EXIT_CODE using sleep recovery mode..."
+        sleep 10000d
+    fi
+
+    echo "Exiting with $_EXIT_CODE..."
+    exit "$_EXIT_CODE"
+}
+
+function exit_init {
+    local _EXIT_CODE="$1"
+    exit_stage "$_EXIT_CODE" "$CP_CAP_RECOVERY_MODE_INIT"
+}
+
+function exit_exec {
+    local _EXIT_CODE="$1"
+    exit_stage "$_EXIT_CODE" "$CP_CAP_RECOVERY_MODE_EXEC"
+}
+
+function exit_term {
+    local _EXIT_CODE="$1"
+    exit_stage "$_EXIT_CODE" "$CP_CAP_RECOVERY_MODE_TERM"
+}
+
+function call_api() {
+    local _HTTP_METHOD="$1"
+    local _API_METHOD="$2"
+    local _HTTP_BODY="$3"
+
+    echo "Calling $API$_API_METHOD..." >> "$LOG_DIR/launch.sh.call_api.log"
+    if [[ "$_HTTP_BODY" ]]; then
+        curl -f -s -k --max-time 30 \
+            -X "$_HTTP_METHOD" \
+            --header 'Accept: application/json' \
+            --header 'Authorization: Bearer '"$API_TOKEN" \
+            --header 'Content-Type: application/json' \
+            --data "$_HTTP_BODY" \
+            "$API$_API_METHOD" \
+            >>"$LOG_DIR/launch.sh.call_api.log" 2>&1
+    else
+        curl -f -s -k --max-time 30 \
+            -X "$_HTTP_METHOD" \
+            --header 'Accept: application/json' \
+            --header 'Authorization: Bearer '"$API_TOKEN" \
+            --header 'Content-Type: application/json' \
+            "$API$_API_METHOD" \
+            >>"$LOG_DIR/launch.sh.call_api.log" 2>&1
+    fi
+}
+
+function tag_run() {
+    local _KEY="$1"
+    local _VALUE="$2"
+    call_api "POST" "run/$RUN_ID/tag" '{
+        "tags": {
+            "'"$_KEY"'": "'"$_VALUE"'"
+        }
+    }'
+}
+
 ######################################################
 
 
@@ -947,6 +1017,15 @@ else
     SINGLE_RUN=false;
 fi
 
+if check_cp_cap CP_CAP_RECOVERY || check_cp_cap RESUMED_RUN ; then
+    export CP_CAP_RECOVERY_MODE_INIT="${CP_CAP_RECOVERY_MODE_INIT:-continue}"
+    export CP_CAP_RECOVERY_MODE_EXEC="${CP_CAP_RECOVERY_MODE_EXEC:-sleep}"
+else
+    export CP_CAP_RECOVERY_MODE_INIT="${CP_CAP_RECOVERY_MODE_INIT:-exit}"
+    export CP_CAP_RECOVERY_MODE_EXEC="${CP_CAP_RECOVERY_MODE_EXEC:-continue}"
+fi
+export CP_CAP_RECOVERY_MODE_TERM="${CP_CAP_RECOVERY_MODE_TERM:-exit}"
+export CP_CAP_RECOVERY_TAG="${CP_CAP_RECOVERY_TAG:-RECOVERED}"
 
 ######################################################
 # Configure Hyperthreading
@@ -1030,7 +1109,7 @@ then
       if [ "$_CP_UPGRADE_RESULT" -ne 0 ]
       then
             echo "[WARN] Packages upgrade done with exit code $_CP_UPGRADE_RESULT, review any issues above"
-            exit "$_DOWNLOAD_RESULT"
+            exit_init "$_DOWNLOAD_RESULT"
       else
             echo "Packages upgrade done"
       fi
@@ -1072,7 +1151,7 @@ if [ ! -f "$CP_PYTHON2_PATH" ]; then
             if [ -z "$CP_PYTHON2_PATH" ]
             then
                   echo "[ERROR] python2 environment not found, exiting."
-                  exit 1
+                  exit_init 1
             fi
       fi
 fi
@@ -1407,7 +1486,7 @@ CP_PIPE_COMMON_ENABLED=${CP_PIPE_COMMON_ENABLED:-"true"}
 if [ "$CP_PIPE_COMMON_ENABLED" == "true" ]; then
       if [ -z "$DISTRIBUTION_URL" ]; then
             echo "[ERROR] Distribution URL is not defined. Exiting"
-            exit 1
+            exit_init 1
       else
             cd $COMMON_REPO_DIR
             # Fixed setuptools version to be compatible with the pipe-common package
@@ -1417,7 +1496,7 @@ if [ "$CP_PIPE_COMMON_ENABLED" == "true" ]; then
             if [ "$_DOWNLOAD_RESULT" -ne 0 ];
             then
                   echo "[ERROR] Main repository download failed. Exiting"
-                  exit "$_DOWNLOAD_RESULT"
+                  exit_init "$_DOWNLOAD_RESULT"
             fi
             _INSTALL_RESULT=0
             tar xf pipe-common.tar.gz
@@ -1426,7 +1505,7 @@ if [ "$CP_PIPE_COMMON_ENABLED" == "true" ]; then
             if [ "$_INSTALL_RESULT" -ne 0 ];
             then
                   echo "[ERROR] Main repository install failed. Exiting"
-                  exit "$_INSTALL_RESULT"
+                  exit_init "$_INSTALL_RESULT"
             fi
             cd -
       fi
@@ -1467,7 +1546,7 @@ if [ "$CP_PIPE_CLI_ENABLED" == "true" ]; then
 
             if [ $? -ne 0 ]; then
                   echo "[ERROR] 'pipe' CLI download failed. Exiting"
-                  exit 1
+                  exit_init 1
             fi
 
             # Clean any known locations, where previous version of the pipe might reside (E.g. committed by the user)
@@ -1507,7 +1586,7 @@ elif [ "$CP_FSBROWSER_ENABLED" == "true" ]; then
       download_file "${DISTRIBUTION_URL}${CP_FSBROWSER_NAME}"
       if [ $? -ne 0 ]; then
             echo "[ERROR] Unable to install FSBrowser"
-            exit 1
+            exit_init 1
       fi
 
       rm -f /bin/fsbrowser
@@ -1543,7 +1622,7 @@ if [ "$CP_GPUSTAT_ENABLED" != "false" ] && check_installed "nvidia-smi"; then
       download_file "${DISTRIBUTION_URL}${CP_GPUSTAT_NAME}"
       if [ $? -ne 0 ]; then
             echo "[ERROR] Unable to download gpustat"
-            exit 1
+            exit_init 1
       fi
 
       CP_GPUSTAT_INSTALL_DIR=${CP_GPUSTAT_INSTALL_DIR:-${CP_USR_BIN}/gpustat-dist}
@@ -1605,7 +1684,7 @@ else
       if [ "$_CLONE_RESULT" -ne 0 ];
       then
             echo "[ERROR] Pipeline repository clone failed. Exiting"
-            exit "$_CLONE_RESULT"
+            exit_init "$_CLONE_RESULT"
       fi
       cd -
 fi
@@ -1705,7 +1784,7 @@ fi
 if [ "$_SETUP_RESULT" -ne 0 ];
 then
     echo "[ERROR] Cluster setup failed. Exiting"
-    exit "$_SETUP_RESULT"
+    exit_init "$_SETUP_RESULT"
 fi
 
 if [ "${OWNER}" ] && [ -d /root/.ssh ]; then
@@ -1781,7 +1860,7 @@ if [ "$CP_DATA_LOCALIZATION_ENABLED" == "true" ]; then
 
             if [ $? -ne 0 ]; then
                   echo "Failed to upload input data"
-                  exit 1
+                  exit_init 1
             fi
             echo
 
@@ -1797,7 +1876,7 @@ if [ "$CP_DATA_LOCALIZATION_ENABLED" == "true" ]; then
                   if [ "$CP_LOCALIZE_FROM_FILES_KEEP_JOB_ON_FAILURE" == "true" ]; then
                         echo "--> It is requested to continue running on config files based localization failure"
                   else
-                        exit 1
+                        exit_init 1
                   fi
             fi
             echo
@@ -2048,31 +2127,36 @@ echo "-"
 if ( check_cp_cap "CP_CAP_SYSTEMD_CONTAINER" || check_cp_cap "CP_CAP_KUBE" ) \
     && check_installed "systemctl" && \
     [ "$CP_OS" == "centos" ] || [ "$CP_OS" == "rocky" ]; then
-      _CONTAINER_DOCKER_ENV_EXPORTING="export container=docker"
-      _IGNORING_CHROOT_ENV_EXPORTING="export SYSTEMD_IGNORE_CHROOT=1"
-      _REMOVING_SYSTEMD_UNIT_PROBLEM_FILES_COMMAND='(cd /lib/systemd/system/sysinit.target.wants/; \
-      for i in *; do \
-        [ $i == systemd-tmpfiles-setup.service ] || rm -f $i; \
-      done); \
-      rm -f /lib/systemd/system/multi-user.target.wants/*;\
-      rm -f /etc/systemd/system/*.wants/*;\
-      rm -f /lib/systemd/system/local-fs.target.wants/*; \
-      rm -f /lib/systemd/system/sockets.target.wants/*udev*; \
-      rm -f /lib/systemd/system/sockets.target.wants/*initctl*; \
-      rm -f /lib/systemd/system/basic.target.wants/*;\
-      rm -f /lib/systemd/system/anaconda.target.wants/*;'
+        _SYSTEMCTL_STATUS=$(systemctl &> /dev/null; $?)
+        if [ "$_SYSTEMCTL_STATUS" -eq 0 ]; then
+            echo "Systemd already active, skipping installation"
+        else
+            _CONTAINER_DOCKER_ENV_EXPORTING="export container=docker"
+            _IGNORING_CHROOT_ENV_EXPORTING="export SYSTEMD_IGNORE_CHROOT=1"
+            _REMOVING_SYSTEMD_UNIT_PROBLEM_FILES_COMMAND='(cd /lib/systemd/system/sysinit.target.wants/; \
+            for i in *; do \
+              [ $i == systemd-tmpfiles-setup.service ] || rm -f $i; \
+            done); \
+            rm -f /lib/systemd/system/multi-user.target.wants/*;\
+            rm -f /etc/systemd/system/*.wants/*;\
+            rm -f /lib/systemd/system/local-fs.target.wants/*; \
+            rm -f /lib/systemd/system/sockets.target.wants/*udev*; \
+            rm -f /lib/systemd/system/sockets.target.wants/*initctl*; \
+            rm -f /lib/systemd/system/basic.target.wants/*;\
+            rm -f /lib/systemd/system/anaconda.target.wants/*;'
 
-      echo $_CONTAINER_DOCKER_ENV_EXPORTING >> /etc/cp_env.sh
-      eval "$_CONTAINER_DOCKER_ENV_EXPORTING"
-      echo $_IGNORING_CHROOT_ENV_EXPORTING >> /etc/cp_env.sh
-      eval "$_IGNORING_CHROOT_ENV_EXPORTING"
-      eval "$_REMOVING_SYSTEMD_UNIT_PROBLEM_FILES_COMMAND"
-      /usr/lib/systemd/systemd --system &
-      
-      # This directory does not exist by default
-      # If it is missing - systemctl will throw "Failed to get D-Bus connection: Operation not permitted"
-      # See: https://serverfault.com/a/925694
-      mkdir /run/systemd/system
+            echo $_CONTAINER_DOCKER_ENV_EXPORTING >> /etc/cp_env.sh
+            eval "$_CONTAINER_DOCKER_ENV_EXPORTING"
+            echo $_IGNORING_CHROOT_ENV_EXPORTING >> /etc/cp_env.sh
+            eval "$_IGNORING_CHROOT_ENV_EXPORTING"
+            eval "$_REMOVING_SYSTEMD_UNIT_PROBLEM_FILES_COMMAND"
+            /usr/lib/systemd/systemd --system &
+
+            # This directory does not exist by default
+            # If it is missing - systemctl will throw "Failed to get D-Bus connection: Operation not permitted"
+            # See: https://serverfault.com/a/925694
+            mkdir /run/systemd/system
+        fi
 else
     echo "Systemd is not requested, skipping installation"
 fi
@@ -2328,6 +2412,10 @@ else
   CP_EXEC_RESULT=$?
 fi
 
+if [ "$CP_EXEC_RESULT" != "0" ]; then
+    exit_exec "$CP_EXEC_RESULT"
+fi
+
 echo "------"
 echo
 ######################################################
@@ -2390,6 +2478,10 @@ else
     rm -Rf $RUN_DIR
 fi
 
-echo "Exiting with $CP_EXEC_RESULT"
-exit "$CP_EXEC_RESULT"
+if [ "$CP_EXEC_RESULT" != "0" ]; then
+    exit_term "$CP_EXEC_RESULT"
+else
+    echo "Exiting with $CP_EXEC_RESULT..."
+    exit "$CP_EXEC_RESULT"
+fi
 ######################################################
